+++ lib/dragonfly_pdf/processors/page.rb
@@ -5,13 +5,11 @@ module DragonflyPdf
   module Processors
     class Page
       def call(content, page_number = 1, _opts = {})
-        @content = content
-        @page_number = page_number
-
+        pdf_properties = DragonflyPdf::Analysers::PdfProperties.new.call(content)
         fail DragonflyPdf::PageNotFound unless pdf_properties[:page_numbers].include?(page_number)
 
-        @content.shell_update(ext: :pdf) do |old_path, new_path|
-          "#{gs_command} -dFirstPage=#{@page_number} -dLastPage=#{@page_number} -o #{new_path} -f #{old_path}"
+        content.shell_update(ext: :pdf) do |old_path, new_path|
+          "#{gs_command} -dFirstPage=#{page_number} -dLastPage=#{page_number} -o #{new_path} -f #{old_path}"
         end
       end
 
@@ -19,10 +17,6 @@ module DragonflyPdf
 
       def gs_command
         'gs -sDEVICE=pdfwrite -dNOPAUSE -dBATCH'
-      end
-
-      def pdf_properties
-        @pdf_properties ||= DragonflyPdf::Analysers::PdfProperties.new.call(@content)
       end
     end
   end