+++ lib/dragonfly_pdf/processors/stamp.rb
@@ -1,13 +1,16 @@
-require 'combine_pdf'
-
 module DragonflyPdf
   module Processors
     class Stamp
       def call(content, stamp_pdf, _options = {})
-        stamp = CombinePDF.parse(stamp_pdf.data).pages[0]
-        result = CombinePDF.parse(content.data)
-        result.pages.each { |page| page << stamp }
-        content.update(result.to_pdf)
+        content.shell_update(ext: :pdf) do |old_path, new_path|
+          "#{pdftk_command} #{old_path} stamp #{stamp_pdf.path} output #{new_path}"
+        end
+      end
+
+      private # =============================================================
+
+      def pdftk_command
+        "pdftk"
       end
     end
   end