+++ test/test_helper.rb
@@ -7,8 +7,6 @@ require 'minitest/spec'
 require 'dragonfly'
 require 'dragonfly_pdf'
 
-require 'pdf-reader'
-
 # ---------------------------------------------------------------------
 
 SAMPLES_DIR = Pathname.new(File.expand_path('../../samples', __FILE__))